MOV : Status into slatekit.common root package
