+++ src/lib/kotlin/slatekit-common/src/main/kotlin/slatekit/common/Status.kt
@@ -1,17 +1,19 @@
-/**
- * <slate_header>
- * url: www.slatekit.com
- * git: www.github.com/code-helix/slatekit
- * org: www.codehelix.co
- * author: Kishore Reddy
- * copyright: 2016 CodeHelix Solutions Inc.
- * license: refer to website and/or github
- * about: A tool-kit, utility library and server-backend
- * mantra: Simplicity above all else
- * </slate_header>
- */
+package slatekit.common
 
-package slatekit.common.status
+/**
+ * Represents the different "states" of a life-cycle
+ */
+sealed class Status(val name:String, val value:Int) {
+    object InActive : Status("InActive", 0)
+    object Starting : Status("Starting", 1)
+    object Idle     : Status("Idle"    , 2)
+    object Running  : Status("Running" , 3)
+    object Paused   : Status("Paused"  , 4)
+    object Stopped  : Status("Stopped" , 5)
+    object Complete : Status("Complete", 6)
+    object Failed   : Status("Failed"  , 7)
+}
+
 
 interface StatusSupport {
