+++ src/lib/kotlin/slatekit-core/src/main/kotlin/slatekit/core/scheduler/Task.kt
@@ -3,8 +3,8 @@ package slatekit.core.scheduler
 import slatekit.common.toId
 import slatekit.core.scheduler.core.ErrorMode
 import slatekit.core.scheduler.core.RunMode
-import slatekit.common.status.Status
-import slatekit.common.status.StatusSupport
+import slatekit.common.Status
+import slatekit.common.StatusSupport
 import java.util.concurrent.atomic.AtomicReference
 
 /**
