+++ src/lib/kotlin/slatekit-workers/src/main/kotlin/slatekit/workers/Worker.kt
@@ -8,7 +8,6 @@ import slatekit.common.metrics.MetricsLite
 import slatekit.common.queues.QueueSource
 import slatekit.common.requests.toResponse
 import slatekit.common.results.ResultCode.NOT_IMPLEMENTED
-import slatekit.common.status.*
 import slatekit.workers.core.*
 import slatekit.workers.core.RunStatus
 import java.util.concurrent.atomic.AtomicReference
