+++ src/lib/kotlin/slatekit-workers/src/main/kotlin/slatekit/workers/core/RunStatus.kt
@@ -14,7 +14,7 @@
 package slatekit.workers.core
 
 import slatekit.common.DateTime
-import slatekit.common.status.Status
+import slatekit.common.Status
 
 data class RunStatus(
     val id: String = "",
