+++ src/lib/kotlin/slatekit-workers/src/main/kotlin/slatekit/workers/core/Stats.kt
@@ -15,7 +15,7 @@ package slatekit.workers.core
 import slatekit.common.DateTime
 import slatekit.common.requests.Response
 import slatekit.common.ResultEx
-import slatekit.common.status.Status
+import slatekit.common.Status
 import slatekit.workers.WorkRequest
 
 /**
